IL: Fix Moms and Babies Eligibilty (BE)
If a household eligible for FamilyCare and Moms and Babies selects an
immigration status filter that excludes FamilyCare, we want to show Moms
and Babies.

Since the immigration filters are applied on the FE, we'll send an API
response with eligibility for both programs and a new
`excludes_programs` field that the FE will use to ensure the program
eligibility is mutually exclusive based on the applied citizenship
filters.

- Add excludes_programs field
- Update program model and admin views
- Update MomsAndBabies calculator to remove FamilyCare eligibility check

https://github.com/MyFriendBen/benefits-api/issues/1085