Fix import program config command
I ran into an issue trying to import TX: EITC on staging due to there being two FederalPovertyLimit objects with a "2025" period: a "THIS_YEAR" (year="THIS_YEAR", period="2025") record and a "2025" record (year="2025", period="2025").

This fix adds a `year` param to the `get` call so that it targets a specific record.

https://linear.app/myfriendben/issue/MFB-428/import-program-config-command-fails-when-program-year-is-this-year